Add some more tests

+++ JCViewOrientations.m
@@ -20,8 +20,8 @@ CGSize JCSizeRestrainedToViewAspectRatio(CGSize size, CGSize constraint, JCViewA
     CGFloat widthRatio = constraint.width / size.width;
     CGFloat heightRatio = constraint.height / size.height;
 
-    CGFloat heightToFit = ceil(size.height * widthRatio);
-    CGFloat widthToFit = ceil(size.width * heightRatio);
+    CGFloat heightToFit = ceilf(size.height * widthRatio);
+    CGFloat widthToFit = ceilf(size.width * heightRatio);
 
     if (aspectRatio == JCViewAspectRatioFill)
     {
